+++ geode-core/src/main/java/org/apache/geode/internal/cache/InternalCacheForClientAccess.java
@@ -1027,7 +1027,7 @@ public class InternalCacheForClientAccess implements InternalCache {
   }
 
   @Override
-  public int getUpTime() {
+  public long getUpTime() {
     return delegate.getUpTime();
   }
